+++ packages/core/src/view/style/edge/Orthogonal.ts
@@ -144,7 +144,11 @@ function getJettySize(state: CellState, isSource: boolean): number {
 /**
  * Implements a local orthogonal router between the given cells.
  *
- * This EdgeStyle is registered under `orthogonalEdgeStyle` in {@link StyleRegistry} when using {@link Graph} or calling {@link registerDefaultEdgeStyles}.
+ * This EdgeStyle is registered under `orthogonalEdgeStyle` in {@link EdgeStyleRegistry} when using {@link Graph} or calling {@link registerDefaultEdgeStyles}.
+ *
+ * **IMPORTANT**: When registering it manually  in {@link EdgeStyleRegistry}, the following metadata must be used:
+ * - handlerKind: 'segment'
+ * - isOrthogonal: true
  *
  * @param state {@link CellState} that represents the edge to be updated.
  * @param sourceScaled {@link CellState} that represents the source terminal.