+++ packages/core/src/view/style/edge/TopToBottom.ts
@@ -25,7 +25,11 @@ import type { EdgeStyleFunction } from '../../../types';
 /**
  * Implements a vertical elbow edge.
  *
- * This EdgeStyle is registered under `topToBottomEdgeStyle` in {@link StyleRegistry} when using {@link Graph} or calling {@link registerDefaultEdgeStyles}.
+ * This EdgeStyle is registered under `topToBottomEdgeStyle` in {@link EdgeStyleRegistry} when using {@link Graph} or calling {@link registerDefaultEdgeStyles}.
+ *
+ * **IMPORTANT**: When registering it manually  in {@link EdgeStyleRegistry}, the following metadata must be used:
+ * - handlerKind: 'elbow'
+ * - isOrthogonal: true
  */
 export const TopToBottom: EdgeStyleFunction = (
   state: CellState,